fixed up dependency and directory trees
